o  changed line 464 which produces a warning on gcc 4.5/Ubuntu    local tests (which show no warning, produce expected output)
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/indexing@116 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ src/mmap.c
@@ -461,7 +461,8 @@ SEXP mmap_extract (SEXP index, SEXP field, SEXP dim, SEXP mmap_obj) {
           memcpy(short_buf, 
                  &(data[(index_p[i]-1)*sizeof(short)]),
                  sizeof(char)*sizeof(short));
-          int_dat[i] = (int)(short)*(short *)(short_buf); 
+          /*int_dat[i] = (int)(short)*(short *)(short_buf); */
+          int_dat[i] = (int)*((short *)short_buf); 
         }
         } else {
         for(i=0;  i < LEN; i++) {
